# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Accounting System",
    "description": "Personal finance tracking with bank statement imports",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/categories": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "List Categories",
        "description": "List all categories (flat list).",
        "operationId": "list_categories_api_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  },
                  "type": "array",
                  "title": "Response List Categories Api Categories Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "categories"
        ],
        "summary": "Create Category",
        "description": "Create a new category.",
        "operationId": "create_category_api_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/categories/{category_id}": {
      "delete": {
        "tags": [
          "categories"
        ],
        "summary": "Delete Category",
        "description": "Delete a category (if no transactions are using it).",
        "operationId": "delete_category_api_categories__category_id__delete",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/budgets": {
      "get": {
        "tags": [
          "budgets"
        ],
        "summary": "List Budgets",
        "operationId": "list_budgets_api_budgets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BudgetResponse"
                  },
                  "type": "array",
                  "title": "Response List Budgets Api Budgets Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "budgets"
        ],
        "summary": "Create Budget",
        "operationId": "create_budget_api_budgets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/budgets/{budget_id}": {
      "patch": {
        "tags": [
          "budgets"
        ],
        "summary": "Update Budget",
        "operationId": "update_budget_api_budgets__budget_id__patch",
        "parameters": [
          {
            "name": "budget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Budget Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "budgets"
        ],
        "summary": "Delete Budget",
        "operationId": "delete_budget_api_budgets__budget_id__delete",
        "parameters": [
          {
            "name": "budget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Budget Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BudgetCreate": {
        "properties": {
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "monthly_limit": {
            "type": "number",
            "title": "Monthly Limit"
          }
        },
        "type": "object",
        "required": [
          "category_id",
          "monthly_limit"
        ],
        "title": "BudgetCreate"
      },
      "BudgetResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "category_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Name"
          },
          "monthly_limit": {
            "type": "number",
            "title": "Monthly Limit"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category_id",
          "monthly_limit",
          "created_at"
        ],
        "title": "BudgetResponse"
      },
...
      "BudgetUpdate": {
        "properties": {
          "monthly_limit": {
            "type": "number",
            "title": "Monthly Limit"
          }
        },
        "type": "object",
        "required": [
          "monthly_limit"
        ],
        "title": "BudgetUpdate"
      },
      "CategoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "is_expense": {
            "type": "boolean",
            "title": "Is Expense",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CategoryCreate",
        "description": "Fields for creating a category."
      },
      "CategoryResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "is_expense": {
            "type": "boolean",
            "title": "Is Expense",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/CategoryResponse"
            },
            "type": "array",
            "title": "Children"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "CategoryResponse",
        "description": "Category response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}